fi_pick_rarest_chunk(): fixed assertion failure.
When attempting to pick the rarest chunk among the set of the offered
range by a remote partial source, we forgot to check that the missing
chunk we want to download was indeed overlapping with the rarest chunk,
during the selection loop.

Later on, the code was assuming that the candidate chunk (the missing
file part we want to download) was indeed overlapping with the rarest
chunk, but it was not always the case.

Hopefully this condition is happening only in rare situations, when we
are downloading from servers having only a partial source to offer.