cq_free(): ensure we can forcefully free events.
At close time, when a callout queue is discarded, we must use the new
ev_forced_free() routine instead of the regular ev_free().

Indeed, the ev_free() routine checks whether the event is no longer part
of the callout queue, which is not true when we destroy the queue!