dq_node_removed(): must free dq object asynchronously.
Finally nailed down this nasty bug, whereby we would happen to remove a
node whilst sending data to it, as part of the DQ processing code.  This
re-entered the DQ code to request freeing of all the DQ objects attached
to that node, and that would cause problems in dq_results_expired(),
where we first send data to a node, then manipulate the dq object.

The symptom was an assertion failure in dquery_check() when calling
dq_install_results_expired() to install the result expiration timer.
That function was precisely added to help capture the situation where
things got corrupted!