socket_tls_upgrade(): synchronous upgrade logged but no longer fatal.
It should not be possible to perform a TLS upgrade synchronously.  If it
happens, something bad must be going on, but this is not our doing.

Therefore, do not panic: simply log that as a critical event (if it
happens regularly, then something about our assumptions will be clearly
proven wrong and we will deal with it). For now abort the upgrading process.